+++ core/components/cronmanager/docs/changelog.md
@@ -5,11 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.2.8] - TBA
+## [1.3.0] - 2022-09-09
+
+### Added
+
+- Settings grid for CronManager settings
+- The cronjobs grid can be sorted by all columns
+
+### Changed
+
+- Move the log panel to a new tab
+- Open a window to view the log entries of a single cronjob
 
 ### Fixed
 
 - Improved fix for showing NaN dates in Safari
+- Purge logs without error deletes all log entries
 
 ## [1.2.7] - 2022-08-15